tests/wpt/web-platform-tests/html/semantics/forms/attributes-common-to-form-controls/dirname-ltr.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Submitting element directionality: the dirname attribute</title>
<link rel="author" title="Denis Ah-Kang" href="mailto:denis@w3.org">
<link rel=help href="https://html.spec.whatwg.org/multipage/#submitting-element-directionality:-the-dirname-attribute">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/dirname.js"></script>
<div id="log"></div>
<form action="resources/dirname-iframe.html" method=get target="iframe">
  <p><label>Comment: <input type=text name="comment" dirname="comment.dir" required></label></p>
  <p><button type=submit>Post Comment</button></p>
</form>
<iframe name="iframe"></iframe>
<script>
  var t = async_test("submit element directionality");
  document.querySelector("input").value = "foobar";
  document.querySelector("button").click();

  onIframeLoadedDone(t, function(params) {
    assert_equals(params.get("comment.dir"), "ltr");
  });
</script>
